+++ crates/nostr/src/key/public_key.rs
@@ -14,6 +14,7 @@ use serde::{Deserialize, Deserializer, Serialize};
 
 use super::Error;
 use crate::nips::nip19::FromBech32;
+use crate::nips::nip21::NostrURI;
 
 /// Public Key
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -42,17 +43,20 @@ impl fmt::Display for PublicKey {
 }
 
 impl PublicKey {
-    /// Try to parse [PublicKey] from `hex` or `bech32`
+    /// Try to parse [PublicKey] from `hex`, `bech32` or [NIP21](https://github.com/nostr-protocol/nips/blob/master/21.md) uri
     pub fn parse<S>(public_key: S) -> Result<Self, Error>
     where
         S: AsRef<str>,
     {
         let public_key: &str = public_key.as_ref();
-        match Self::from_hex(public_key) {
+        match Self::from_nostr_uri(public_key) {
             Ok(public_key) => Ok(public_key),
-            Err(_) => match Self::from_bech32(public_key) {
+            Err(_) => match Self::from_hex(public_key) {
                 Ok(public_key) => Ok(public_key),
-                Err(_) => Err(Error::InvalidPublicKey),
+                Err(_) => match Self::from_bech32(public_key) {
+                    Ok(public_key) => Ok(public_key),
+                    Err(_) => Err(Error::InvalidPublicKey),
+                },
             },
         }
     }
@@ -65,8 +69,6 @@ impl PublicKey {
     }
 
     /// Parse [PublicKey] from `hex` string
-    ///
-    /// Use `PublicKey::from_str` to try to parse it from `hex` or `bech32`
     pub fn from_hex<S>(hex: S) -> Result<Self, Error>
     where
         S: AsRef<str>,
@@ -90,7 +92,7 @@ impl PublicKey {
 impl FromStr for PublicKey {
     type Err = Error;
 
-    /// Try to parse [PublicKey] from `hex` or `bech32`
+    /// Try to parse [PublicKey] from `hex`, `bech32` or [NIP21](https://github.com/nostr-protocol/nips/blob/master/21.md) uri
     fn from_str(public_key: &str) -> Result<Self, Self::Err> {
         Self::parse(public_key)
     }
@@ -112,5 +114,68 @@ impl<'de> Deserialize<'de> for PublicKey {
     {
         let hex: String = String::deserialize(deserializer)?;
         Self::parse(hex).map_err(serde::de::Error::custom)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    pub fn test_public_key_parse() {
+        let public_key = PublicKey::parse(
+            "nostr:npub14f8usejl26twx0dhuxjh9cas7keav9vr0v8nvtwtrjqx3vycc76qqh9nsy",
+        )
+        .unwrap();
+        assert_eq!(
+            public_key.to_hex(),
+            "aa4fc8665f5696e33db7e1a572e3b0f5b3d615837b0f362dcb1c8068b098c7b4"
+        );
+    }
+}
+
+#[cfg(bench)]
+mod benches {
+    use test::{black_box, Bencher};
+
+    use super::*;
+
+    const NIP21_URI: &str = "nostr:npub14f8usejl26twx0dhuxjh9cas7keav9vr0v8nvtwtrjqx3vycc76qqh9nsy";
+    const HEX: &str = "aa4fc8665f5696e33db7e1a572e3b0f5b3d615837b0f362dcb1c8068b098c7b4";
+    const BECH32: &str = "npub14f8usejl26twx0dhuxjh9cas7keav9vr0v8nvtwtrjqx3vycc76qqh9nsy";
+
+    #[bench]
+    pub fn parse_public_key_nip21_uri(bh: &mut Bencher) {
+        bh.iter(|| {
+            black_box(PublicKey::parse(NIP21_URI)).unwrap();
+        });
+    }
+
+    #[bench]
+    pub fn parse_public_key_hex(bh: &mut Bencher) {
+        bh.iter(|| {
+            black_box(PublicKey::parse(HEX)).unwrap();
+        });
+    }
+
+    #[bench]
+    pub fn public_key_from_hex(bh: &mut Bencher) {
+        bh.iter(|| {
+            black_box(PublicKey::from_hex(HEX)).unwrap();
+        });
+    }
+
+    #[bench]
+    pub fn parse_public_key_bech32(bh: &mut Bencher) {
+        bh.iter(|| {
+            black_box(PublicKey::parse(BECH32)).unwrap();
+        });
+    }
+
+    #[bench]
+    pub fn public_key_from_bech32(bh: &mut Bencher) {
+        bh.iter(|| {
+            black_box(PublicKey::from_bech32(BECH32)).unwrap();
+        });
     }
 }